refactor: move client related configs to client.configs.ts file
