--- src/client/constants/code-editor.ts
@@ -1,16 +1,0 @@
-export const CODE_EDITOR_TEMPLATES_ARRAY = ['test-runner-js', 'react'] as const;
-
-export const CODE_EDITOR_TEMPLATES = {
-  TEST_RUNNER_JS: 'test-runner-js',
-  REACT: 'react',
-};
-
-export const CODE_EDITOR_BOILERPLATE_CATEGORIES_ARRAY = [
-  'course',
-  'blog',
-] as const;
-
-export const CODE_EDITOR_BOILERPLATE_CATEGORIES = {
-  COURSE: 'course',
-  BLOG: 'blog',
-};
