--- src/client/constants/site-settings.ts
@@ -1,14 +1,0 @@
-export const SITE_TITLE =
-  'Codealoy - Programming and Web Development Learning Platform in Bangla';
-
-export const SITE_NAME = 'Codealoy';
-
-export const SITE_NAME_LOWERCASE = 'codealoy';
-
-export const BASE_URL = 'https://www.codealoy.com';
-
-export const TWITTER_USER_NAME = 'codealoy';
-
-export const COLORS = {
-  PRIMARY: '#5628ee',
-};
