+++ src/client/types/code-editor-types.d.ts
@@ -1,7 +1,7 @@
 import {
   CODE_EDITOR_BOILERPLATE_CATEGORIES_ARRAY,
   CODE_EDITOR_TEMPLATES_ARRAY,
-} from '~/client/constants/code-editor';
+} from '~/configs/client.configs';
 
 export type CodeEditorTempates = (typeof CODE_EDITOR_TEMPLATES_ARRAY)[number];
