+++ src/client/utils/getCustomBoilerplate.ts
@@ -1,11 +1,12 @@
 import { SandpackProviderProps } from '@codesandbox/sandpack-react';
+
 import { courseBoilerplates } from '~/client/boilerplates/courses';
-import { CODE_EDITOR_BOILERPLATE_CATEGORIES } from '~/client/constants/code-editor';
 import {
   CodeEditorBoilerplateCategoires,
   CodeEditorBoilerplates,
 } from '~/client/types/code-editor-types';
-import { getCourseAndModuleFromBoilerplate } from './getCourseAndModuleFromBoilerplate';
+import { getCourseAndModuleFromBoilerplate } from '~/client/utils/getCourseAndModuleFromBoilerplate';
+import { CODE_EDITOR_BOILERPLATE_CATEGORIES } from '~/configs/client.configs';
 
 export interface GetCustomBoilerplateParams {
   boilerplate: CodeEditorBoilerplates;
