+++ src/configs/client.configs.ts
@@ -1,0 +1,38 @@
+// Site Configs
+export const SITE_TITLE =
+  'Codealoy - Programming and Web Development Learning Platform in Bangla';
+
+export const SITE_NAME = 'Codealoy';
+
+export const SITE_NAME_LOWERCASE = 'codealoy';
+
+export const BASE_URL = 'https://www.codealoy.com';
+
+export const TWITTER_USER_NAME = 'codealoy';
+
+export const SITE_COLORS = {
+  PRIMARY: '#5628ee',
+};
+
+export const SUPPORT_EMAIL = 'mailto:support@codealoy.com';
+
+// Theme Configs
+export const DEFAULT_THEME = 'dark';
+
+// Code Editor Configs
+export const CODE_EDITOR_TEMPLATES_ARRAY = ['test-runner-js', 'react'] as const;
+
+export const CODE_EDITOR_TEMPLATES = {
+  TEST_RUNNER_JS: 'test-runner-js',
+  REACT: 'react',
+};
+
+export const CODE_EDITOR_BOILERPLATE_CATEGORIES_ARRAY = [
+  'course',
+  'blog',
+] as const;
+
+export const CODE_EDITOR_BOILERPLATE_CATEGORIES = {
+  COURSE: 'course',
+  BLOG: 'blog',
+};
