+++ src/pages/index.tsx
@@ -1,16 +1,17 @@
-import { type NextPage } from 'next';
 import Head from 'next/head';
 
-import { CTA } from '~/client/components/home/CTA';
+import { type NextPage } from 'next';
+
 // import { FAQ } from '~/components/home/FAQ';
 import { Divider } from '~/client/components/Divider';
+import { CTA } from '~/client/components/home/CTA';
 import { FeatureGrid } from '~/client/components/home/FeatureGrid';
 import { Hero } from '~/client/components/home/Hero';
 import { HighlightedFeatures } from '~/client/components/home/HighlightedFeatures';
 import { LatestBlogPosts } from '~/client/components/home/LatestBlogPosts';
 import { LatestCourses } from '~/client/components/home/LatestCourses';
 import { Team } from '~/client/components/home/Team';
-import { SITE_TITLE } from '~/client/constants/site-settings';
+import { SITE_TITLE } from '~/configs/client.configs';
 
 const IndexPage: NextPage = () => {
   return (
